fix: improve delayed initialization of input modes
Input modes are not loaded immediately on startup. There is a very
short delay. During this time all input is just dropped. This is to
make sure that the input buffer is flushed of any rubbish left over
from the previous application.

This commit fixes delayed initialization in cases where an input mode
is switched immediately on startup. For example when the home screen
loads.