feat: improve scrollbar behavior
Clicking and dragging inside the scrollbar grip will now preserve the mouse click offset
and not jump to the center position on click.

Clicking outside the scrollbar grip will now jump as though the center of the grip was
dragged to the click position.

This is more inline with typical GUI scrollbar behaviors.